feat: 섹션 헤더를 날짜로 바꾸고 기존 텍스트를 시간으로 변경
